fix:[ENG-2072] address PR review comments
- Fix content truncation in explain mode (200→2000 to match default)
- Standardize score precision to .toFixed(4) in both display modes
- Move providerTypeToLabel from domain types to CLI renderer layer
- Remove dead `?? chalk.dim` fallback from colorLabel (exhaustive Record)
- Truncate enrichmentKeywords to first 5 words per entry instead of
  passing full content blobs
- Add note to --explain flag description about JSON mode behavior
