+++ src/agent/core/domain/swarm/types.ts
@@ -126,32 +126,6 @@ export type QueryResult = {
   score: number
 }
 
-/**
- * Human-readable source label for a provider, used in CLI output.
- */
-export function providerTypeToLabel(type: ProviderType, id: string): string {
-  switch (type) {
-    case 'byterover': { return 'context-tree'
-    }
-
-    case 'gbrain': { return 'gbrain'
-    }
-
-    case 'hindsight': { return 'hindsight'
-    }
-
-    case 'honcho': { return 'honcho'
-    }
-
-    case 'local-markdown': {
-      const name = id.split(':')[1] ?? 'files'
-      return `notes:${name}`
-    }
-
-    case 'obsidian': { return 'obsidian'
-    }
-  }
-}
 
 /**
  * Result from storing a memory entry.
